requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
google-generativeai
pypdf
python-dotenv
